requirements.txt
playwright>=1.59
openai>=2.36
python-dotenv>=1.2
rich>=15.0
pydantic>=2.13
pytest>=9.0
mcp[cli]
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
httpx>=0.27,<1.0
